// src/models/LocStats.ts
export interface LocStats {
  tsFiles: number;
  tsLines: number;
  tsComments: number;
  jsFiles: number;
  jsLines: number;
  jsComments: number;
  cssFiles: number;
  cssLines: number;
  cssComments: number;
  scssFiles: number;
  scssLines: number;
  scssComments: number;
  htmlFiles: number;
  htmlLines: number;
  htmlComments: number;
  pumlFiles: number;
  pumlLines: number;
  pumlComments: number;
  mdFiles: number;
  mdLines: number;
  mdComments: number;
}